requirements.txt
pillow
qrcode==8.2
